changed add_numbers
We generalized the function to receive float parameters.

+++ add_numbers.cpp
@@ -1,11 +1,11 @@
 #include<iostream>
 using namespace std;
 
-int add(int n1, int n2){
+float add(float n1, float n2){
 	return n1+n2;
 }
 int main(){
-	int num1, num2;
+	float num1, num2;
 	cin>>num1>>num2;
 	cout<<add(num1,num2);
 }
